Remove compile warnings
Mostly by changing to explicit casts.  But check these: unsigned ;ong
to double IS a narrowing cast.  Also one warning remains:

src/FMC_FE3.cpp:654:61: warning: format ‘%ld’ expects argument of type
    ‘long int’, but argument 3 has type ‘long int*’ [-Wformat=]
    fprintf(fout, "\tFuelKey: %ld\n", MxDesc.FuelKey[i][j]);

    long  FuelKey   [eC_Sta] [NUM_FUEL_SIZES] [256];

So this is a 3d array but the code in FMC_FE3 only refers to the first 2
dimensions, so yes, this is a pointer, not a long.

+++ src/fsxwmech.cpp
@@ -839,7 +839,7 @@ double Mechanix::spreadrate(long slope, double windspd, int fuel)
      if(ndead>0)
           ndead=4;
 
-     double nclas[2]={ndead,nlive};  // # of dead & live fuel classes
+     double nclas[2]={static_cast<double>(ndead),static_cast<double>(nlive)};  // # of dead & live fuel classes
      double xmext[2]={newfuel.xmext, 0};
 
      double load[4][2]=			     // tons per acre, later converted to lb/ft2
@@ -867,11 +867,11 @@ double Mechanix::spreadrate(long slope, double windspd, int fuel)
      //-------------------------------------------------------------------------
 
 	double sav[4][2]=			     // 1/ft
-	{	{newfuel.sav1, newfuel.savlh},
-		{109.0, newfuel.savlw},
-		{30.0, 0.0},
-		{newfuel.savlh, 0.0},
-	};
+         {	{static_cast<double>(newfuel.sav1), static_cast<double>(newfuel.savlh)},
+            {109.0, static_cast<double>(newfuel.savlw)},
+            {30.0, 0.0},
+            {static_cast<double>(newfuel.savlh), 0.0},
+         };
 
 	double heat[4][2]=
      {    {newfuel.heatd, newfuel.heatl},
